Fixed warnings that have functional change and actually fixes some bugs WD2793.c   - Fixed bug in wd2793PeekDataReg() where wd->dataReg was not updated and the non-updated wd->dataReg was returned RomMapper.c   - Fixed bug in romMapperGuessRom() that caused mapper type detection to never detect any other than ROM_STANDARD. romMapperKonamiWordPro.c   - Fixed bug in Write() that caused control register to always be written instead of only for address 0xbfff romMapperSCCplus.c   - Fixed bug in loadState() that cased sccMode to be set to SCC_COMPATIBLE for any other than SCC_PLUS types. So only these two types were restored correctly. Fmopl.c   - Fixed bug in init_timetables() that initialized some arrays out of bounds. This could cause undefined behavior (might or might not be an issue in practice depending on lots of factors like compiler used). CRTC6845.c   - Fixed bug in crtc6845Reset() where 'registers' and 'cursor' was not entirely cleared

+++ source/Memory/RomMapper.c
@@ -315,10 +315,10 @@ RomType romMapperGuessRom(const void *buffer, int size, int guess, char* extende
 
     /* Count occurences of characteristic addresses */
     for (i = 0; i < size - 2; i++) {
-        UInt8 value = romData[i] + ((int)romData[i + 1] << 8) + ((int)romData[i + 2] << 16);
+        UInt32 value = romData[i] + ((UInt32)romData[i + 1] << 8) + ((UInt32)romData[i + 2] << 16);
 
         switch(value) {
-        case 0x400032: 
+        case 0x400032:
             counters[3]++;
             break;
